Create xHx-autoservers.js
an early (starting) hack script that is a continuation of early-hack-script tutorial and includes automatic purchased servers upgrades from 2GB to max.
it will break if interrupted before end of cycle. recommended to use 'serverDelete.js' in that case.
runtime errors about missing files on beginning/first run can be ignored with no problems --- it is from using this same file as a re-start file as well and user WILL have the missing files.